+++ truck-meshalgo/tests/tessellation/triangulation.rs
@@ -17,7 +17,7 @@ fn read_jsons() -> Vec<Vec<u8>> {
 #[test]
 fn solid_is_closed() {
     for (i, json) in read_jsons().into_iter().enumerate() {
-        let solid = Solid::extract(serde_json::from_reader(json.as_slice()).unwrap()).unwrap();
+        let solid: Solid = serde_json::from_reader(json.as_slice()).unwrap();
         let mut poly = solid.triangulation(0.02).unwrap().to_polygon();
         poly.put_together_same_attrs()
             .remove_degenerate_faces()
@@ -34,7 +34,7 @@ fn solid_is_closed() {
 #[test]
 fn compare_occt_mesh() {
     let jsons = read_jsons();
-    let solid = Solid::extract(serde_json::from_slice(jsons[2].as_slice()).unwrap()).unwrap();
+    let solid: Solid = serde_json::from_slice(jsons[2].as_slice()).unwrap();
     let res = solid.triangulation(0.01).unwrap().to_polygon();
     let path = concat!(dir!(), "../obj/by_occt.obj");
     let ans = obj::read(std::fs::read(path).unwrap().as_slice()).unwrap();
@@ -45,6 +45,6 @@ fn compare_occt_mesh() {
 #[test]
 fn large_number_meshing() {
     let json = std::fs::read(concat!(dir!(), "large-torus.json")).unwrap();
-    let torus = Solid::extract(serde_json::from_slice(json.as_slice()).unwrap()).unwrap();
+    let torus: Solid = serde_json::from_slice(json.as_slice()).unwrap();
     let _ = torus.triangulation(1.0).unwrap().to_polygon();
 }